feat: Improved 'env apply' and 'env destroy' usability

+++ plugins/kubectl/util/delete-secrets.sh
@@ -4,13 +4,13 @@
 : "${taito_project:?}"
 : "${taito_env:?}"
 
-echo "- ${taito_project}-${taito_env}-bucket"
+# TODO obsolete?
 (
   ${taito_setv:?}
   kubectl delete secret "${taito_project}-${taito_env}-bucket" 2> /dev/null
 )
 
-echo "- ${taito_project}-${taito_env}-basic-auth"
+# TODO obsolete?
 (
   ${taito_setv:?}
   kubectl delete secret "${taito_project}-${taito_env}-basic-auth" 2> /dev/null
@@ -21,11 +21,12 @@ secret_names=(${taito_secret_names})
 for secret_name in "${secret_names[@]}"
 do
   . "${taito_cli_path}/util/secret-by-index.sh"
-  (
-    ${taito_setv:?}
-    kubectl delete secret "${secret_name}" --namespace="${secret_namespace}"
-  )
-  echo "- ${secret_name} deleted"
+  if [[ ${secret_method:?} != "read/"* ]]; then
+    (
+      ${taito_setv:?}
+      kubectl delete secret "${secret_name}" --namespace="${secret_namespace:?}"
+    )
+  fi
   secret_index=$((${secret_index}+1))
 done
